requirements.txt
numpy>=1.19.0
torch>=1.7.0
torchvision>=0.8.0
transformers>=4.5.0
pillow>=8.0.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
flask>=2.0.0
requests>=2.25.0
tqdm>=4.60.0
diffusers>=0.16.0